Time: 0 ms (100%), Space: 113.7 MB (85.53%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.cpp
@@ -1,14 +1,14 @@
 class Solution {
 public:
     double findMaxAverage(vector<int>& nums, int k) {
-        long long sum = 0;
+        int sum = 0;
         int left = 0;
         int right = k - 1;
 
         for(int i= 0; i < k; i++){
             sum += nums[i];
         }
-        long long max_sum = sum;
+        int max_sum = sum;
 
         while(right + 1 < nums.size()){
             sum -= nums[left]; 
@@ -18,6 +18,6 @@ public:
             max_sum = sum > max_sum? sum : max_sum;
         }
 
-        return ((double)max_sum / k);
+        return (double)max_sum / k;
     }
 };
